Add splitting of large polygons.
New command line option -S/--split.

When they are too large (more than N points) they are split in half.  They are
split horizontally or vertically, in whichever direction they are larger. This
is done recursively.

If there is a geometry error when splitting, the polygon is left alone and not
split. Unfortunately this means we need some more checks to fix broken polygons
before we can do the splitting. That needs to be done in a later commit.

This commit also contains code to output rings to the error_lines table that
have the wrong orientation, ie. islands where the land is on the right instead
of left side of the way tagged as coastline.