Simplifying and splitting polygon code.
For some use cases the simplification and splitting code in osmcoastline isn't
the right thing. Added here are some SQL scripts and documentation how
simplification and splitting on tile boundaries can be done with PostGIS. This
is not really coastline specific, other large polygons could be treated in the
same way.