Add extra check for missing positions.
When the input data doesn't contain all the nodes needed in the coastline ways,
the output could be corrupt. This commits adds a test to check that all
positions are there and stops processing if nodes are missing. If the debug
option is set ids of missing nodes are reported.