Add preliminary support for comparing coastlines between runs.
This commit adds an option to the osmcoastline program to write out coastline
segments to a file. Two files from different runs can then be compared with the
new osmcoastline_segments program. Currently it only supports writing out
changes to stdout and creating a shapefile with the changes. In the future
different uses are possible.